fix(app): resolve remaining coderabbit findings
